Move upgrade tests to examples
