+++ examples/go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/aws/aws-sdk-go-v2 v1.30.5
 	github.com/aws/aws-sdk-go-v2/config v1.27.33
 	github.com/aws/aws-sdk-go-v2/service/eks v1.48.4
-	github.com/pulumi/providertest v0.0.11
+	github.com/pulumi/providertest v0.0.15
 	github.com/pulumi/pulumi-aws/sdk/v6 v6.41.0
 	github.com/pulumi/pulumi-eks/sdk/v3 v3.0.0-alpha.6
 	github.com/pulumi/pulumi/pkg/v3 v3.129.0
